Refactor; move `IntMatrix` to separate file
